Add active flag to default texts form
